app/requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
click==8.1.7
Django==5.0.2
django-ninja==1.1.0
gunicorn==21.2.0
h11==0.14.0
packaging==23.2
psycopg==3.1.18
pydantic==2.6.1
pydantic_core==2.16.2
sqlparse==0.4.4
typing_extensions==4.9.0
uvicorn==0.27.1
